third_party/WebKit/LayoutTests/crypto/subtle/pbkdf2-deriveBits-empty-password-failure-expected.txt

Tests importing empty password for PBKDF2

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "PBKDF2"
PASS key.usages.join(',') is "deriveKey,deriveBits"
error is: OperationError: The password for PBKDF2 must not be empty.
PASS successfullyParsed is true

TEST COMPLETE
